django/project/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.2
django-ckeditor-5==0.2.17
django-cors-headers==4.6.0
djangorestframework==3.15.2
django-unfold==0.49.1
idna==3.10
msal==1.31.1
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.8.2
django-extensions==3.2.3
Werkzeug==3.0.3
pyOpenSSL==24.3.0
